# src/express/tsoa-generated/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Facility": {
				"description": "テナントに属する施設",
				"properties": {
					"tenantId": {
						"type": "string"
					},
					"id": {
						"type": "string",
						"description": "データの一意性を判定するために必ず定義する"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"tenant": {
						"$ref": "#/components/schemas/Tenant"
					}
				},
				"required": [
					"tenantId",
					"name",
					"createdAt",
					"updatedAt",
					"tenant"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tenant": {
				"description": "テナント(お客様組織)",
				"properties": {
					"tenantId": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"description": "コンソールなどでも表示される、エンドユーザーが認識しやすい名称\n例: \"株式会社しらかわ\""
					},
					"code": {
						"type": "string",
						"description": "URLの一部などに利用される、開発者が認識しやすい名称\n例: \"my-company\""
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"facilities": {
						"items": {
							"$ref": "#/components/schemas/Facility"
						},
						"type": "array"
					}
				},
				"required": [
					"tenantId",
					"name",
					"code",
					"createdAt",
					"updatedAt",
					"facilities"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Facility.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"RegisterFacilityReq": {
				"$ref": "#/components/schemas/Pick_Facility.name_"
			},
			"Pick_Tenant.name-or-code_": {
				"properties": {
					"name": {
						"type": "string",
						"description": "コンソールなどでも表示される、エンドユーザーが認識しやすい名称\n例: \"株式会社しらかわ\""
					},
					"code": {
						"type": "string",
						"description": "URLの一部などに利用される、開発者が認識しやすい名称\n例: \"my-company\""
					}
				},
				"required": [
					"name",
					"code"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"RegisterTenantReq": {
				"$ref": "#/components/schemas/Pick_Tenant.name-or-code_"
			}
		},
		"securitySchemes": {
			"TENANT": {
				"type": "apiKey",
				"name": "x-tenant-id",
				"in": "header",
				"description": "とりあえずテナントIDをつっこむ。本格的なセキュリティはこんど考える"
			}
		}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
		"/facilities": {
			"post": {
				"operationId": "RegisterFacilityGMO",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Facility"
								}
							}
						}
					}
				},
				"tags": [
					"02: Facility"
				],
				"security": [
					{
						"API": [
							"TENANT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterFacilityReq"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListFacilities",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Facility"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"02: Facility"
				],
				"security": [
					{
						"API": [
							"TENANT"
						]
					}
				],
				"parameters": []
			}
		},
		"/samples/get/{path}": {
			"get": {
				"operationId": "SampleGet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"99: Sample"
				],
				"security": [
					{
						"NONE": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/samples/error": {
			"get": {
				"operationId": "SampleError",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"99: Sample"
				],
				"security": [
					{
						"NONE": []
					}
				],
				"parameters": []
			}
		},
		"/samples/busy": {
			"get": {
				"operationId": "SampleBusy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"99: Sample"
				],
				"security": [
					{
						"NONE": []
					}
				],
				"parameters": []
			}
		},
		"/tenants": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tenant"
								}
							}
						}
					}
				},
				"description": "新たな Tenant を登録します",
				"tags": [
					"01: Tenant"
				],
				"security": [
					{
						"API": [
							"SUPER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "名前、コード",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterTenantReq"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Tenant"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"01: Tenant"
				],
				"security": [
					{
						"API": [
							"SUPER"
						]
					}
				],
				"parameters": []
			}
		},
		"/tenants/{tenantId}": {
			"get": {
				"operationId": "Find",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tenant"
								}
							}
						}
					}
				},
				"tags": [
					"01: Tenant"
				],
				"security": [
					{
						"API": [
							"SUPER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tenants/{tenantId}/name": {
			"patch": {
				"operationId": "UpdateName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tenant"
								}
							}
						}
					}
				},
				"tags": [
					"01: Tenant"
				],
				"security": [
					{
						"API": [
							"TENANT"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/tenants/{tenantId}/code": {
			"patch": {
				"operationId": "UpdateCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tenant"
								}
							}
						}
					}
				},
				"tags": [
					"01: Tenant"
				],
				"security": [
					{
						"API": [
							"TENANT"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"code": {
										"type": "string"
									}
								},
								"required": [
									"code"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8080/api/v1"
		}
	]
}